+++ src/Controller/Admin/PointOfInterestController.php
@@ -3,13 +3,14 @@
 namespace App\Controller\Admin;
 
 use App\Entity\PointOfInterest;
-use EasyCorp\Bundle\EasyAdminBundle\Config\Asset;
+use App\Field\VichFileField;
+use App\Field\VichImageField;
+use App\Service\EasyAdminHelper;
+use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use EasyCorp\Bundle\EasyAdminBundle\Field\DateField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\IdField;
-use EasyCorp\Bundle\EasyAdminBundle\Field\ImageField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
-use EasyCorp\Bundle\EasyAdminBundle\Form\Type\FileUploadType;
 use Symfony\Component\Translation\TranslatableMessage;
 
 class PointOfInterestController extends AbstractCrudController
@@ -21,21 +22,37 @@ class PointOfInterestController extends AbstractCrudController
 
     public function configureFields(string $pageName): iterable
     {
-        return [
-            IdField::new('id')->hideOnForm(),
-            TextField::new('name')->setRequired(true)
-                ->setHelp(new TranslatableMessage('The name of the point of interest', [], 'admin')),
-            TextField::new('latitude')->setRequired(true)
-                ->setHelp(new TranslatableMessage('The latitude of the interest point', [], 'admin')),
-            TextField::new('longitude')->setRequired(true)
-            ->setHelp(new TranslatableMessage('The longitude of the interest point', [], 'admin')),
-            TextField::new('subtitles')->setRequired(true)
-            ->setHelp(new TranslatableMessage('A text version of the podcast, for people with hearing disabilities.', [], 'admin')),
-            ImageField::new('image')->setRequired(true)->setUploadDir('/public/points-of-interest')->hideWhenUpdating(),
-            TextField::new('podcast')->setRequired(true)->setFormType(FileUploadType::class)->addJsFiles(Asset::fromEasyAdminAssetPackage('field-file-upload.js'))->hideWhenUpdating(),
-            IdField::new('id')->hideOnForm(),
-            DateField::new('createdAt')->hideOnForm(),
-            DateField::new('updatedAt')->hideOnForm(),
-        ];
+        yield IdField::new('id')->hideOnForm();
+        yield TextField::new('name');
+        yield TextField::new('subtitles')->setRequired(true)
+        ->setHelp(new TranslatableMessage('A text version of the podcast, for people with hearing disabilities.', [], 'admin'));
+        yield TextField::new('latitude')->setRequired(true)
+        ->setHelp(new TranslatableMessage('The latitude of the interest point.', [], 'admin'));
+        yield TextField::new('longitude')->setRequired(true)
+        ->setHelp(new TranslatableMessage('The longitude of the interest point.', [], 'admin'));
+
+        $context = $this->getContext();
+        if (in_array($pageName, [Crud::PAGE_NEW, Crud::PAGE_EDIT], true) && null !== $context) {
+            $entity = $context->getEntity()->getInstance();
+            assert($entity instanceof PointOfInterest);
+
+            $imageAttr = EasyAdminHelper::getFileInputAttributes($entity, 'imageFile');
+            yield VichImageField::new('imageFile')
+                ->onlyOnForms()
+                ->setFormTypeOption('allow_delete', false)
+                ->setFormTypeOption('attr', $imageAttr);
+
+            $podcastAttr = EasyAdminHelper::getFileInputAttributes($entity, 'podcastFile');
+            yield VichFileField::new('podcastFile')
+                ->onlyOnForms()
+                ->setFormTypeOption('allow_delete', false)
+                ->setFormTypeOption('attr', $podcastAttr);
+        } else {
+            yield VichImageField::new('image');
+            yield VichFileField::new('podcast');
+        }
+
+        yield DateField::new('createdAt')->hideOnForm();
+        yield DateField::new('updatedAt')->hideOnForm();
     }
 }