+++ src/Controller/Admin/RouteController.php
@@ -3,11 +3,13 @@
 namespace App\Controller\Admin;
 
 use App\Entity\Route;
+use App\Field\VichImageField;
+use App\Service\EasyAdminHelper;
+use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use EasyCorp\Bundle\EasyAdminBundle\Field\AssociationField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\DateField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\IdField;
-use EasyCorp\Bundle\EasyAdminBundle\Field\ImageField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
 use Symfony\Component\Translation\TranslatableMessage;
 
@@ -20,21 +22,32 @@ class RouteController extends AbstractCrudController
 
     public function configureFields(string $pageName): iterable
     {
-        return [
-            IdField::new('id')->hideOnForm(),
-            TextField::new('name')
-            ->setHelp(new TranslatableMessage('The name of the route', [], 'admin')),
-            TextField::new('description'),
-            TextField::new('distance')
-            ->setHelp(new TranslatableMessage('The distance should be how far the route is with all points of interests included, e.g. "840m"', [], 'admin')),
-            ImageField::new('image')->setUploadDir('/public/routes')->hideWhenUpdating(),
-            IdField::new('id')->hideOnForm(),
-            AssociationField::new('tags')->hideOnIndex()->setRequired(true)->setFormTypeOption('by_reference', false)
-                ->setHelp(new TranslatableMessage('Tags are used in the frontend to organize the routes. If the route is not connected to a tag, it will not be displayed in the frotnend', [], 'admin')),
-            AssociationField::new('pointsOfInterest')->hideOnIndex()->setRequired(true)
-            ->setHelp(new TranslatableMessage('Connect points of interest to this podwalk', [], 'admin')),
-            DateField::new('createdAt')->hideOnForm(),
-            DateField::new('updatedAt')->hideOnForm(),
-        ];
+        yield IdField::new('id')->hideOnForm();
+        yield TextField::new('name');
+        yield TextField::new('description');
+        yield TextField::new('distance')
+            ->setHelp(new TranslatableMessage('The distance should be how far the route is with all points of interests included, e.g. "840m"', [], 'admin'));
+
+        $context = $this->getContext();
+        if (in_array($pageName, [Crud::PAGE_NEW, Crud::PAGE_EDIT], true) && null !== $context) {
+            $entity = $context->getEntity()->getInstance();
+            assert($entity instanceof Route);
+            $attr = EasyAdminHelper::getFileInputAttributes($entity, 'imageFile');
+
+            yield VichImageField::new('imageFile')
+                ->onlyOnForms()
+                ->setFormTypeOption('allow_delete', false)
+                ->setFormTypeOption('attr', $attr);
+        } else {
+            yield VichImageField::new('image');
+        }
+
+        yield IdField::new('id')->hideOnForm();
+        yield AssociationField::new('tags')->hideOnIndex()->setRequired(true)->setFormTypeOption('by_reference', false)
+        ->setHelp(new TranslatableMessage('Tags are used in the frontend to organize the routes.', [], 'admin'));
+        yield AssociationField::new('pointsOfInterest')->hideOnIndex()->setRequired(true)
+            ->setHelp(new TranslatableMessage('Connect points of interest to this podwalk.', [], 'admin'));
+        yield DateField::new('createdAt')->hideOnForm();
+        yield DateField::new('updatedAt')->hideOnForm();
     }
 }