+++ src/Entity/Route.php
@@ -10,34 +10,58 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 use Gedmo\Timestampable\Traits\TimestampableEntity;
+use Symfony\Component\HttpFoundation\File\File;
+use Symfony\Component\Serializer\Annotation\Groups;
+use Symfony\Component\Serializer\Annotation\SerializedName;
+use Symfony\Component\Validator\Constraints as Assert;
+use Vich\UploaderBundle\Mapping\Annotation as Vich;
 
 #[ORM\Entity(repositoryClass: RouteRepository::class)]
 #[ApiResource(
+    normalizationContext: ['groups' => ['read']],
     operations: [
         new Get(),
         new GetCollection(),
-    ],
-    security: "is_granted('ROLE_API_USER')"
+    ]
 )]
+#[Vich\Uploadable]
 class Route
 {
     use TimestampableEntity;
+
     #[ORM\Id]
     #[ORM\GeneratedValue]
     #[ORM\Column]
     private ?int $id = null;
 
     #[ORM\Column(length: 255)]
+    #[Groups('read')]
     private ?string $name = null;
 
     #[ORM\Column(length: 255)]
+    #[Groups('read')]
     private ?string $description = null;
 
     #[ORM\Column(length: 255)]
+    #[Groups('read')]
     private ?string $distance = null;
 
     #[ORM\Column(length: 255)]
     private ?string $image = null;
+
+    #[Vich\UploadableField(mapping: 'uploads', fileNameProperty: 'image')]
+    #[Assert\File(
+        mimeTypes: [
+            'image/jpeg',
+            'image/png',
+        ],
+        maxSize: '1m'
+    )]
+    private ?File $imageFile = null;
+
+    #[Groups('read')]
+    #[SerializedName('image')]
+    public ?string $imageUrl = null;
 
     #[ORM\ManyToMany(targetEntity: PointOfInterest::class, inversedBy: 'routes')]
     private Collection $pointsOfInterest;
@@ -99,10 +123,10 @@ class Route
 
     public function getImage(): ?string
     {
-        return '/uploads/images/'.$this->image;
+        return $this->image;
     }
 
-    public function setImage(string $image): static
+    public function setImage(?string $image): static
     {
         $this->image = $image;
 
@@ -158,5 +182,30 @@ class Route
         }
 
         return $this;
+    }
+
+    /**
+     * If manually uploading a file (i.e. not using Symfony Form) ensure an instance
+     * of 'UploadedFile' is injected into this setter to trigger the update. If this
+     * bundle's configuration parameter 'inject_on_load' is set to 'true' this setter
+     * must be able to accept an instance of 'File' as the bundle will inject one here
+     * during Doctrine hydration.
+     *
+     * @param File|\Symfony\Component\HttpFoundation\File\UploadedFile|null $imageFile
+     */
+    public function setImageFile(File $imageFile = null): void
+    {
+        $this->imageFile = $imageFile;
+
+        if (null !== $imageFile) {
+            // It is required that at least one field changes if you are using doctrine
+            // otherwise the event listeners won't be called and the file is lost
+            $this->updatedAt = new \DateTime();
+        }
+    }
+
+    public function getImageFile(): ?File
+    {
+        return $this->imageFile;
     }
 }