+++ src/Field/VichFileField.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace App\Field;
+
+namespace App\Field;
+
+use EasyCorp\Bundle\EasyAdminBundle\Contracts\Field\FieldInterface;
+use EasyCorp\Bundle\EasyAdminBundle\Field\FieldTrait;
+use Vich\UploaderBundle\Form\Type\VichFileType;
+
+final class VichFileField implements FieldInterface
+{
+    use FieldTrait;
+
+    public static function new(string $propertyName, string $label = null): self
+    {
+        return (new self())
+            ->setProperty($propertyName)
+            ->setLabel($label)
+            // this template is used in 'index' and 'detail' pages
+            ->setTemplatePath('admin/field/vich_file.html.twig')
+            // this is used in 'edit' and 'new' pages to edit the field contents
+            // you can use your own form types too
+            ->setFormType(VichFileType::class)
+            ->addCssClass('field-vich-file')
+        ;
+    }
+}