+++ CHANGELOG.md
@@ -17,3 +17,4 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 - Add Route
 - Add Tags
 - Add fixtures
+- Image upload improved